Reorganize module code structure
