+++ torch/csrc/api/include/torch/nn/modules/modules.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <torch/nn/modules/batchnorm.h>
+#include <torch/nn/modules/containers.h>
+#include <torch/nn/modules/conv.h>
+#include <torch/nn/modules/dropout.h>
+#include <torch/nn/modules/embedding.h>
+#include <torch/nn/modules/functional.h>
+#include <torch/nn/modules/linear.h>
+#include <torch/nn/modules/modules.h>
+#include <torch/nn/modules/rnn.h>
