+++ torch/csrc/api/include/torch/torch.h
@@ -1,4 +1,7 @@
 #pragma once
-#include "torch/containers.h"
+
+#include <torch/nn/module.h>
+#include <torch/nn/modules/modules.h>
+
 #include "torch/optimizers.h"
 #include "torch/serialization.h"
