+++ torch/csrc/api/src/optimizers.cpp
@@ -14,7 +14,7 @@ void OptimizerImpl::zero_grad() {
   }
 }
 
-void OptimizerImpl::set_model(Container model) {
+void OptimizerImpl::set_model(std::shared_ptr<nn::Module> model) {
   model_ = model;
 }
